Refactor Email Verification link
Refactored the email verification link generation and validation to prevent validation of accounts by a user that does not own the email: If the "evil" user found out the id of the newly generated user, we could recreate the validation link. This refactor prevents it by using encripting instead of hashing the email.