Resolves #22: "A guest should be able to create an account story" (#34)
* wip

* Create accounts

Adds account creation functionality, closing #22

* remove useless code

* remove cached views

* remove cached views

* Adjust test due to random failures

Clean code.
Corrected namespace.
Updated LinkStoreTest to use 'https://www.google.com' instead of a faker generated email since it seems in some cases faker is generating urls that fail the 'active_url' rule (due to a failing dns query).

* Updated changelog

Updated changelog.
Improved account store validation rules.

* Improvements

As per https://github.com/laravel-portugal/api/pull/34#pullrequestreview-501834067.

* Refactor Email Verification link

Refactored the email verification link generation and validation to prevent validation of accounts by a user that does not own the email: If the "evil" user found out the id of the newly generated user, we could recreate the validation link. This refactor prevents it by using encripting instead of hashing the email.

* Improvements/refactoring

* Remove duplicate Changelog entry

* Re-adding mail.conf

Exception happens when this file is missing.
This is not catched by the tests since no email is actually sent. We assert the notification is "sent", but in testing, it is not really email or even pushed into a queue for that matter.

* Revert Email verify refactor

* fix(.env): add missing environment variables related to the mailer service

Co-authored-by: José Postiga <jpostiga@infraspeak.com>